Cleaned up the multipart response DSL.
